+++ Public/Realtime/Set-RealtimeTranscriptionSessionConfiguration.ps1
@@ -1,20 +1,12 @@
-function Set-RealtimeSessionConfiguration {
+function Set-RealtimeTranscriptionSessionConfiguration {
     [CmdletBinding()]
     param (
         [Parameter()]
         [string]$EventId,
 
         [Parameter()]
-        [AllowEmptyString()]
-        [string]$Instructions,
-
-        [Parameter()]
         [ValidateSet('text', 'audio')]
         [string[]]$Modalities = @('text'),
-
-        [Parameter()]
-        [Completions('alloy', 'ash', 'ballad', 'coral', 'echo', 'fable', 'nova', 'onyx', 'sage', 'shimmer', 'verse')]
-        [string][LowerCaseTransformation()]$Voice,
 
         [Parameter()]
         [Completions(
@@ -25,20 +17,9 @@ function Set-RealtimeSessionConfiguration {
         [string][LowerCaseTransformation()]$InputAudioFormat,
 
         [Parameter()]
-        [Completions(
-            'pcm16',
-            'g711_ulaw',
-            'g711_alaw'
-        )]
-        [string][LowerCaseTransformation()]$OutputAudioFormat,
-
-        [Parameter()]
         [AllowEmptyString()]
         [Completions('near_field', 'far_field', 'none')]
         [string][LowerCaseTransformation()]$InputAudioNoiseReductionType,
-
-        [Parameter()]
-        [bool]$EnableInputAudioTranscription,
 
         [Parameter()]
         [Completions('gpt-4o-transcribe', 'gpt-4o-mini-transcribe', 'whisper-1')]
@@ -71,31 +52,21 @@ function Set-RealtimeSessionConfiguration {
         [Parameter()]
         [uint16]$TurnDetectionSilenceDuration,
 
+        # Not available for transcription sessions.
         [Parameter()]
         [bool]$CreateResponseOnTurnEnd = $true,
 
+        # Not available for transcription sessions.
         [Parameter()]
         [bool]$InterruptResponse = $true,
 
         [Parameter()]
-        [ValidateNotNullOrEmpty()]
-        [System.Collections.IDictionary[]]$Tools,
-
-        [Parameter()]
-        [Completions('none', 'auto', 'required')]
-        [string]$ToolChoice,
-
-        [Parameter()]
-        [ValidateRange(0.6, 1.2)]
-        [float]$Temperature,
-
-        [Parameter()]
-        [ValidateRange(-1, 4096)]
-        [int]$MaxResponseOutputTokens = -1
+        [AllowEmptyCollection()]
+        [string[]]$Include
     )
 
     begin {
-        $MessageObject = @{type = 'session.update'; session = @{} }
+        $MessageObject = @{type = 'transcription_session.update'; session = @{} }
     }
 
     process {
@@ -108,14 +79,8 @@ function Set-RealtimeSessionConfiguration {
         if ($PSBoundParameters.ContainsKey('Modalities')) {
             $MessageObject.session.modalities = $Modalities
         }
-        if ($PSBoundParameters.ContainsKey('Voice')) {
-            $MessageObject.session.voice = $Voice
-        }
         if ($PSBoundParameters.ContainsKey('InputAudioFormat')) {
             $MessageObject.session.input_audio_format = $InputAudioFormat
-        }
-        if ($PSBoundParameters.ContainsKey('OutputAudioFormat')) {
-            $MessageObject.session.output_audio_format = $OutputAudioFormat
         }
 
         if ($PSBoundParameters.ContainsKey('InputAudioNoiseReductionType')) {
@@ -128,23 +93,16 @@ function Set-RealtimeSessionConfiguration {
             }
         }
 
-        if ($PSBoundParameters.ContainsKey('EnableInputAudioTranscription')) {
-            if (-not $EnableInputAudioTranscription) {
-                $MessageObject.session.input_audio_transcription = $null
-            }
-            else {
-                $InputAudioTranscriptionParam = @{
-                    model = $InputAudioTranscriptionModel
-                }
-                if ($PSBoundParameters.ContainsKey('InputAudioTranscriptionLanguage')) {
-                    $InputAudioTranscriptionParam.language = $InputAudioTranscriptionLanguage
-                }
-                if ($PSBoundParameters.ContainsKey('InputAudioTranscriptionPrompt')) {
-                    $InputAudioTranscriptionParam.prompt = $InputAudioTranscriptionPrompt
-                }
-                $MessageObject.session.input_audio_transcription = $InputAudioTranscriptionParam
-            }
+        $InputAudioTranscriptionParam = @{
+            model = $InputAudioTranscriptionModel
         }
+        if ($PSBoundParameters.ContainsKey('InputAudioTranscriptionLanguage')) {
+            $InputAudioTranscriptionParam.language = $InputAudioTranscriptionLanguage
+        }
+        if ($PSBoundParameters.ContainsKey('InputAudioTranscriptionPrompt')) {
+            $InputAudioTranscriptionParam.prompt = $InputAudioTranscriptionPrompt
+        }
+        $MessageObject.session.input_audio_transcription = $InputAudioTranscriptionParam
 
         if ($PSBoundParameters.ContainsKey('EnableTurnDetection')) {
             if (-not $EnableTurnDetection) {
@@ -173,23 +131,8 @@ function Set-RealtimeSessionConfiguration {
             }
         }
 
-        if ($PSBoundParameters.ContainsKey('Temperature')) {
-            $MessageObject.session.temperature = $Temperature
-        }
-        if ($PSBoundParameters.ContainsKey('MaxResponseOutputTokens')) {
-            if ($MaxResponseOutputTokens -lt 0) {
-                $MessageObject.session.max_response_output_tokens = 'inf'
-            }
-            else {
-                $MessageObject.session.max_response_output_tokens = $MaxResponseOutputTokens
-            }
-        }
-
-        if ($PSBoundParameters.ContainsKey('Tools')) {
-            $MessageObject.session.tools = $Tools
-        }
-        if ($PSBoundParameters.ContainsKey('ToolChoice')) {
-            $MessageObject.session.tool_choice = $ToolChoice
+        if ($PSBoundParameters.ContainsKey('Include')) {
+            $MessageObject.session.include = $Include
         }
 
         PSOpenAI\Send-RealtimeSessionEvent -Message ($MessageObject | ConvertTo-Json -Depth 10)